little tidy of file organisation ready for future work
